Fix: create logs/ui directory before writing log file

+++ src/test/java/eu/europa/eudi/utils/factory/MobileDeviceLogger.java
@@ -39,6 +39,10 @@ public class MobileDeviceLogger {
                     featureName + "_" + scenarioNumber + ".log"
             );
 
+            if (!logDir.exists() && !logDir.mkdirs()) {
+                throw new IOException("Cannot create log directory: " + logDir);
+            }
+
             if (logFile.exists() && !logFile.delete()) {
                 throw new IOException("Cannot delete old log file: " + logFile);
             }
